+++ src/main/java/org/openphc/cce/insights/web/dto/AdoptionKpiDto.java
@@ -11,10 +11,12 @@ public class AdoptionKpiDto {
     private String facilityName;
     /** Validated baseline from facility (set by programme staff). */
     private long expectedVisitsPerDay;
-    /** Average daily distinct reporters over the selected period (or today when no range). */
-    private long actualVisitsPerDay;
+    /** Average daily distinct reporters over the selected period (or today when no range).
+     *  Float so small averages (e.g. 15 visits over 90 days = 0.17) are not rounded to 0. */
+    private double actualVisitsPerDay;
     /** actual / expected × 100 (period total when a date range is supplied). */
     private double adoptionRate;
-    /** Average daily expected − actual; positive = under-reporting. */
-    private long reportingGapPerDay;
+    /** Average daily expected − actual; positive = under-reporting. Float for the same
+     *  reason as {@link #actualVisitsPerDay}. */
+    private double reportingGapPerDay;
 }
